Correct pa11y config file format
Pa11y wants request headers to be a simple key-value dictionary,
while Scrapy was outputting a dictionary where the values were
lists of strings. This commit corrects this, and also adds a test
to detect when <title> elements mismatch between pa11y and scrapy,
which is how we identified this bug in the first place.