Specify login credentials
This commit allows the person calling pa11ycrawler to specify an
email and password to use when logging into the Open edX
installation. If the email and password aren't specified, pa11ycrawler
will attempt to create a user via the "auto auth" feature, and
save the email and password for that user.

In addition, if (when) the crawler gets logged out, it can now
automatically log back in using the same email and password. If it
obtained an email and password from the "auto auth" feature, it
will save those credentials and use them to log back in
automatically.